EMP & Wires refactor (#19401)
* EMP Refactor

* EVERYBODY gets wires

* Shield gen

* fix

* Fix those

* Changes

* ,

* kill

---------

+++ code/modules/shieldgen/directional_shield.dm
@@ -227,6 +227,9 @@
 		. += "Its shield matrix is at [round( (shield_health / max_shield_health) * 100, 0.01)]% strength."
 
 /obj/item/shield_projector/emp_act(severity, recursive)
+	. = ..()
+	if (. & EMP_PROTECT_SELF)
+		return
 	adjust_health(-max_shield_health / severity) // A strong EMP will kill the shield instantly, but weaker ones won't on the first hit.
 
 // Subtypes
